seminar_repos/03_data_fun/06_cats_rare_treasures/01_be_integration_testing/db/connection.py
from pg8000.native import Connection
from dotenv import load_dotenv
import os

load_dotenv()

DB_USER = os.environ["DB_USER"]
DB_PASS = os.environ["DB_PASS"]


def create_connection():
    conn = Connection(DB_USER, database="nc_games", password=DB_PASS)
    return conn


def close_connection(conn):
    conn.close()
